+++ sycl/include/sycl/types.hpp
@@ -827,7 +827,7 @@ public:
   void dump() {
 #ifndef __SYCL_DEVICE_ONLY__
     for (int I = 0; I < NumElements; ++I) {
-      printf("  %d: %d\n", I, getValue(I));
+      printf("  %d: %lld\n", I, (long long int)getValue(I));
     }
     printf("\n");
     fflush(stdout);
